Make header block class
