+++ src/main/java/hw2/HomePage.java
@@ -5,6 +5,8 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
+import ru.yandex.qatools.htmlelements.loader.decorator.HtmlElementDecorator;
+import ru.yandex.qatools.htmlelements.loader.decorator.HtmlElementLocatorFactory;
 
 import java.util.List;
 
@@ -14,28 +16,12 @@ public class HomePage {
      */
     public WebDriver driver;
 
+    Header header;
+
     public HomePage(WebDriver driver) {
-        PageFactory.initElements(driver, this);
+        PageFactory.initElements(new HtmlElementDecorator(new HtmlElementLocatorFactory(driver)), this);
         this.driver = driver;
     }
-
-    @FindBy(id = "name")
-    private WebElement loginField;
-
-    @FindBy(id = "password")
-    private WebElement passwdField;
-
-    @FindBy(id = "user-icon")
-    private WebElement loginPasswdMenu;
-
-    @FindBy(id = "login-button")
-    private WebElement loginBtn;
-
-    @FindBy(id = "user-name")
-    private WebElement name;
-
-    @FindBy(css = "ul.nav>li>a")
-    private List<WebElement> navigationTop;
 
     @FindBy(css = "div.benefit-icon")
     private List<WebElement> images;
@@ -61,48 +47,16 @@ public class HomePage {
     @FindBy(tagName = "footer")
     private WebElement footer;
 
-    @FindBy(css = "ul.nav .dropdown")
-    private WebElement serviceTop;
 
     @FindBy(css = "li.menu-title[index=\"3\"]")
     private WebElement serviceLeft;
 
-    /**
-     * метод для ввода логина
-     */
-    public void inputLogin(String login) {
-        loginField.sendKeys(login);
-    }
-
-    /**
-     * метод для ввода пароля
-     */
-    public void inputPasswd(String passwd) {
-        passwdField.sendKeys(passwd);
-    }
-
-    /**
-     * метод для осуществления нажатия кнопки входа в аккаунт
-     */
-    public void clickLoginBtn() {
-        loginBtn.click();
-    }
-
-    /**
-     * метод для осуществления открытия меню ввода логина и пароля
-     */
-    public void openMenu() {
-        loginPasswdMenu.click();
-    }
 
     /**
      * метод для входа в аккаунт
      */
     public void signInWithoutClear(String login, String passwd) {
-        loginPasswdMenu.click();
-        loginField.sendKeys(login);
-        passwdField.sendKeys(passwd);
-        loginBtn.click();
+        header.signInWithoutClear(login, passwd);
     }
 
     /**
@@ -110,20 +64,6 @@ public class HomePage {
      */
     public String getTitle() {
         return driver.getTitle();
-    }
-
-    /**
-     * метод для чтения имени пользователя
-     */
-    public String getName() {
-        return name.getAttribute("innerText");
-    }
-
-    /**
-     * метод для нахождения 4-х навигационных элементов в хедере
-     */
-    public List<WebElement> getHeaderSectionNavigation() {
-        return navigationTop;
     }
 
     /**
@@ -208,20 +148,6 @@ public class HomePage {
     }
 
     /**
-     * метод для получения элемента Service из хедера
-     */
-    public WebElement getServiceTop() {
-        return serviceTop;
-    }
-
-    /**
-     * метод для получения выпадающего списка Service из хедера
-     */
-    public List<WebElement> getServiceListTop() {
-        return getServiceTop().findElements(By.cssSelector(".dropdown-menu li"));
-    }
-
-    /**
      * метод для получения элемента Service из левого блока
      */
     public WebElement getServiceLeft() {
@@ -235,12 +161,4 @@ public class HomePage {
         return getServiceLeft().findElements(By.className("li"));
     }
 
-    /**
-     * метод для перехода на  страницу serviceListElement в Service в хедере
-     */
-    public void goToServiceElement(String serviceListElement) {
-        getServiceTop().click();
-        //getServiceTop().findElement(By.cssSelector("a[href=\"different-elements.html\"]")).click();
-        getServiceTop().findElement(By.linkText(serviceListElement)).click();
-    }
 }
